Basic support for inserting into arbitrary place in list
Now we have an ability, at the db level, to "append" into the middle of a list, instead of just inserting at the beginning.
I didn't test this feature, since i think i'm going to rework lists to make them more manageable. I'm making this commit
as a sort of backup in case that doesn't work, I can go back and not lose this work.