expr-id is not necessarily a list
It seems there'd been some confusion about both body-id and expr-id. body-id
should always be a list-header, as we have (lambda (a b c) one two three).
expr-id is not necessarily a list as we can have (define a b) or
(define a '(one two three)) or (define f (lambda (a b c) one two three))
but not (define (a b c) one two three)