Fixed a bug when replacing a define expr
the bug was where you would have the define's expr be a list, and you'd
type 'a', and it would try to get the cdr_id of the define. This fixes
that bug, so that 'a' and 'i' do nothing if you are on the body of a
define, no matter whether it's a list or not. You can still do 'A' and
'I' on a list in this case, and it works right