split code out into 3 distinct files
previously all code was in 'first-attempt.rkt'. Now we have 'main-thingy.rkt'
for most of the stuff, 'misc.rkt' for common utils, and 'sql-db.rkt' for the db.
currently everything seems to work. Code is not properly factored yet. sql-db
just exports all identifiers raw. But at least this proves that deps are
one-directional