Large refactor / complete rewrite of database
We now have a sql implementation of the db interface. The new impl has
radically changed the logic. a lot of major refactors were done and
db idioms changed a lot. Close enough to being a rewrite not to
attempt to exhaust the large number of differences. This diff is a new
leaf and probably represents how the logic should have been written in
the first place, had i known then what i know now