Access to a lambda's body is now protected
Previously the lambda api just had a get-body-list method which exposed
a veme:db-list%%, which you could do anything with, including unassign!!
We never want the body list to be unassign!!'d, so we had to replace
get-body-list with get-body (which returns a scheme list rather than a
db list handle) and insert/remove methods. Now the lambda is fully in
charge of its own body, and to the caller the body list looks "inlined"
in the lambda, unlike the impl where it's a separate object.

We also had to add some more logic into get-parent, so that the parent
of a lambda body item is the lambda, not the list