scheme-base-model
this was a starter attempt at abstracting the db model so there would be
a db model and a more generic model, one that could be used for in-memory
data like a copied subtree. I've abandoned it for now - for the time being
the middle layer is provided by the the entity model, and i won't worry
much about copy/paste or undo/redo until we get there. However, it may be
useful in the future, so i'm going to commit it in this commit and delete
it in the next one, to preserve what little progress i made while keeping
the working tree clean of distractions