fixed transpilation and handle storage bugs
Apparently racket is satan and won't just hash my damn struct-object. Maybe
init-field would do the trick but i don't want to add more things that the
oop model has to implement. So locs are marshalled back and forth between
id&col for the sake of the handles* hash table.

also fixed a bunch of misc bugs in transpile.rkt