Major perf improvements
perf is still horrendous, and probably will be as long as we use the
current gui. But it's a lot better than it was. The main improvement
is that now we don't have to do a query on every single table every
single time we call get-table, which is invoked for most queries.
The gui is slow, but apparently db queries are even slower so we must
be careful about how much we're doing them going forward