update gui commands
revived all the former commands post refactor, and added some new stuff.
This 100% completes the refactor - but of course there's more refactoring
down the road. Ents need to be refactored - i'll probably get to that when
i roll a new gui, that way the ents can be built with a sane gui in mind