implement optional param front end
all of this is very kludgy and bad. writing proper ent code is enough
of a hassle that i'm going to pivot and rewrite the gui before modeling
the oop. So for now we'll deal with some ugliness at the front. But at
least i was able to use the front end to casually test the back end