another partial commit - saving the item->event-handler pattern
around around we go on the refactor merry go round. i had the pattern
of passing a ui:item% a item->event-handler sort of thing. This seemed
to promote flexibility and not require too much init args, at the
expense of just being a fundamentally weird way of reimplementing
inheritance. At this point it seems that it's hard to reasonably factor
the code doing this. Probably we should not reinvent inheritance, adn
just have ui:item% have a get-event-handler method or some god damn thing.
idk. this is insane.

anyways, the current tree state is fairly advanced and about 2/3 feature
complete. so it's a good idea to maek a partial commit here and save it
before going off and changing everything again for the tenth time