Partial commit
ok just one somewhat arbitrary place to commit at. This is was at the branching
point of deciding how to deal with the problem where switching switching
between opt and req'd params meant unassigning and reassigning a bunch of
handles, which would then leave the reference slots with invalid handles.
Possible fixes include:
1) rewriting the middle layer and ui again, so that the ui is purely
functional, and selection works by selecting a handle rather than a ui, and
no other state is supported really.
2) trying in some way to find all slots and update them with their new handles
3) supporting the change between opt and req'd params in the db, allowing us
to do this operation without invalidating any handles, and for the time being
not supporting the deletion of any referable that is referenced by any
non-descendants

i am probably going with option 3)
option 1) is the most work and i've already rewritten this thing too many times.
option 2) doesn't really seem to work if we want to support modules and stuff in
the future
option 3) is kind of lame but it is easy to do and should work pretty smoothly.
Supporting insta-delete is something we can do in the future after we have
passed some major milestones and have good reason to rewrite the ent/gui system
once again