PARTIAL: saving another attempt at making ent:module% not be an ent:typical-list%
this time, ent:module% is basically similar to ent:typical-list%, but can't
subclass it, as ent:module% ui-list needs to override the child slot event
handler. But that's not actually right, since it's not the child slot which
controls navigation - it's the ui item itself, hidden away in
ent:module-lambda% or such. So really ent:module% should just be a vanilla
ent:typical-list% . But at least save this attempt to the contrary