Adds translate.rkt
translate.rkt is a heuristic program that converts a racket file into
a zinal db. It will read a conforming racket file and create a module
in the specified db that has logic roughly equivalent to the racket file.
A human has to go and manually fix any mistakes, set up requires and
public defs properly, fill in TODOs, and some other stuff as necessary.
We can augment the translator later to do a better job, leaving less manual
cleanup.