add assertions
in the zinal code, assert is a macro defined in misc.rkt . but zinal
doesn't support macros. so in order to self host we need zinal to have
an 'assert' primitive. this seems pretty reasonable, since some other
languages (e.g. java) have assert as a primitive. In the future, we
may augment it to be a general exception-throwing primitive, with
different possible behavior, and then it would be easy to replace
line numbers in the stack trace with something appropriate to zinal.