OOP backend
Interface and implementation of all oop stuff in the database. Doesn't include
transpile, translate, or front end, and there's still a few loose ends that
need to be tied up, so this diff is really a partial in a sequence of diffs for
OOP, but sort-example.db is proof that this diff doesn't completely ruin
everything, and it's already very large and complex, so i'm committing at this
point. Loose ends can be tied up in the near future, and tweaks will undoubtedly
be made as transpile/translate/front-end are fleshed out.

This commit also does some somewhat irrelevant things. The most important one
is it fixes a bug where get-table-mod* was dependent on hash table orderings.
Due to the bug, example .db files in previous commits no longer work for their
own commit. The bug probably triggered after updating to the latest version of
racket (6.8) - so it may be possible to use those old .db files by downgrading
to racket 6.5 . This commit fixes the bug for this and all future commits by
sorting ID-TABLES .