Add oop stuff to translate.rkt
translation is very basic and doesn't bother getting super classes right, or
the methods of method invokations. this stuff must be polished up and augmented
later. Also, the translation is barely tested, since it's hard to test anything
without gui support. Also, translation of is-a? isn't done, since is-a? is
about to be removed from the backend