delete is-a?%%
originally i thought that having zinal:db:reference%% for interfaces would
be too unideal; they would have exactly one use (as the second argument of
an is-a?), they wouldn't be used in places where it'd kind of make sense to
use them (e.g. get-direct-interfaces still returns zinal:db:interface%% ,
not zinal:db:interface-ref%%), and most of the time they'd just add unnecessary
options in the front end referable selection list when the user is selecting
a referable to create a reference to. So i created a primitive is-a? node to
take care of the precisely one purpose for interface references. This was
awkward and unideal for its own reasons, but at the time it seemed the better
choice. After realizing that this would prevent using curry with is-a? , and
that it would prevent passing an interface reference into something like
assert-is* , i decided that interface references were the lesser evil,
deleting the is-a? primitive and creating a reference node that points to
interfaces. This isn't a technical issue since referables don't have to be
nodes.