Fully support the publicity of zinal:db:define-class%%
apparently, i had some of the things necessary for a zinal:db:define-class%%
to be publicly exposed if the direct child of a module, but not all of them.
So basically this diff is a bug fix for some things i forgot to do to take
care of that case.