hack to fix weird gui focus race condition
After upgrading ubuntu to 16.04.1, i started having this problem where most,
but not all of the time, the second dialog in a row would not automatically
have focus. E.g., if you create a new node, first you'll have a dialog to
decide which type, and if you choose, say, reference or number, another dialog
pops up so you can decide the value. The latter dialog would usually but not
always show up without focus, which is really annoying because you are forced
to use the clicker to give it focus. I think the problem is due to the weird
code i wrote in choice-dialog% or whatever to set a value when closing. I think
what happens is that the first dialog isn't 100% finished "closing" before the
second one pops up - as the first finishes closing, it regains focus. But I
have no idea what i should be waiting on in order to ensure the first dialog
doesn't steal focus in its death throes. I also don't know whether this is the
fault of my code or of the racket gui impl. In light of my ignorance, i've made
a dirty hack to fix the problem. oh well. all i need is for it to be passable,
not good