permit gui and translate to search through TODOs
in order for translate.rkt to recognize as many references as possible,
it's essentially necessary to do a second pass through the db, finding and
converting TODOs that were meant to refer to an older sibling or aunt. It's
also expedient for the UI to have a "find next/previous unassigned" feature,
so that the long tail of post-translation TODOs can be quickly found and
fixed manually. This diff implements all this stuff. this diff also implements
a general purpose tree search/iteration algo in misc.rkt