megacommit
a big diff with a lot of related and unrelated changes. Some big ones:
- added ui-styles.rkt - now it is easier to modify look and feel. There was
  also some important practical reason for this but i forgot
- added some features to ents, including method search
- various conformization stuff
- compiled code has a custom error-display-handler that will
  s/zinal:_id-n/short-desc/g
- major fixes/improvements to transpile and translate
- db-util now takes on more stuff

at this point i'm too lazy to split everything properly, but that's not the
only problem. the ui-style change and the db-util change cause unrelated changes
on the same line - afaict there is not a reasonable way to split out changes to
a particular line. In a projectional environment there's no such things as
"lines", so it's easy and natural for diffs to be split out distinct conceptual
units. This is one of those things that the text-based environment is ill-suited
for (at best) and where i believe substantial strides can be made if the
medium for logic expression is structued rather than text (at least as far as
the user is concerned)