add a zinal.db that is a working version of zinal expressed in a zinal db
zinal.db is a demonstration of zinal's ability to self-host

zinal.db is a db that compiles to a version of zinal. it is not equivalent
to any particular revision of zinal in the git repo, but it's pretty close
to the current revision. it was produced by using translate.rkt , after
which a lot of manual fixing had to be done. I also made hacks to ents.rkt
and sql-db.rkt to make the manual fixing-up easier. If you open zinal.db
with zinal, you can make any edits you like to the db, then you can compile,
and after about 15-20 minutes, you will have a (mostly) standalone zinal
executable that reflects any changes you made and which has much better
start-up time than calling `racket main.rkt`